pyproject.toml
[project]
name = "arch-mvc"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.128.1,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "passlib (==1.7.4)",
    "bcrypt (==4.0.1)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
